chore(ci): Prevent duplicate GitHub Actions runs
Previously, workflows used this trigger:

```yaml
on: [push, pull_request]
```

This means run on every push and pull request, to any branch.
This would lead to duplicate runs when creating a pull request, as the push to the branch would trigger the tests, then the pull request would trigger them again.

This PR migrates workflows so the `push` event is only triggered on the `main` branch.